optimize models for registration and editing user. move identical information from this models into one class (UserBasicInfoDTO)
