+++ src/main/java/org/registrator/community/dto/AddressDTO.java
@@ -12,7 +12,7 @@ public class AddressDTO implements Serializable {
 	public static final String ONLY_LITERALS = "[а-яіїєА-ЯІЇЄa-zA-Z,\\s,\\.,\\-]+";
 	public static final String ONLY_DIGITS = "[0-9]+";
 
-	@Pattern(regexp = "[(?=.*[0-9])]*", message = "Поле повинне містити тільки цифри")
+	@Pattern(regexp = "[[0-9])]*", message = "{msg.registration.onlydigits}")
 	private String postcode;
 
 	private String region;
