+++ src/main/java/org/registrator/community/dto/UserDTO.java
@@ -4,196 +4,90 @@ import java.io.Serializable;
 import java.util.List;
 
 import javax.validation.Valid;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Email;
-import org.hibernate.validator.constraints.NotEmpty;
 import org.registrator.community.dto.JSON.ResourceNumberDTOJSON;
 
-public class UserDTO implements Serializable {
+public class UserDTO extends UserBasicInfoDTO implements Serializable {
 
-	private static final long serialVersionUID = 1L;
-	@NotEmpty
-	@Size(min = 1, max = 30, message = "Ім\'я повинно містити від {min} до {max} символів")
-	@Pattern(regexp = AddressDTO.ONLY_LITERALS, message = "Некоректне введення")
-	private String firstName;
+    private static final long serialVersionUID = 1L;
 
-	@NotEmpty
-	@Size(min = 1, max = 30, message = "Прізвище повинне містити від {min} до {max} символів")
-	@Pattern(regexp = AddressDTO.ONLY_LITERALS, message = "Некоректне введення")
-	private String lastName;
+    private String login;
+    
+    private String role;
 
-	@Pattern(regexp = "[(?=.*[а-яіїєА-ЯІЇЄa-zA-Z,\\s,\\.,\\-]).{1,30}]*", message = "Некоректне введення")
-	private String middleName;
+    private String status;
 
-	private String role;
+    @Valid
+    private ResourceNumberDTOJSON resourceNumberDTOJSON;
 
-	private String login;
+    private WillDocumentDTO willDocument;
+    private List<String> otherDocuments;
 
-	@NotEmpty
-	@Email(message = "Введіть коректну адресу")
-	private String email;
+    public UserDTO() {
 
-	private String status;
+    }
 
-	@Valid
-	private AddressDTO address;
+    public UserDTO(String firstName, String lastName, String middleName, String role, String login, String email,
+            String status, AddressDTO address, PassportDTO passport) {
+        super(firstName, lastName, middleName, email, address, passport);
+        this.login = login;
+        this.role = role;
+        this.status = status;
+    }
 
-	@Valid
-	private PassportDTO passport;
+    public UserDTO(String firstName, String lastName, String middleName, String role, String login, String email,
+            String status, AddressDTO address, PassportDTO passport, String territorialCommunity,
+            ResourceNumberDTOJSON resourceNumberDTOJSON) {
+        this(firstName, lastName, middleName, role, login, email, status, address, passport);
+        this.resourceNumberDTOJSON = resourceNumberDTOJSON;
+    }  
 
-	@Valid
-	private ResourceNumberDTOJSON resourceNumberDTOJSON;
-	
+    public String getLogin() {
+        return login;
+    }
 
-	private String territorialCommunity;
+    public void setLogin(String login) {
+        this.login = login;
+    }
 
-	private WillDocumentDTO willDocument;
-	private List<String> otherDocuments;
+    public String getRole() {
+        return role;
+    }
 
-	public UserDTO(String firstName, String lastName, String middleName, String role, String login, String email,
-			String status, AddressDTO address, PassportDTO passport) {
-		this.firstName = firstName;
-		this.lastName = lastName;
-		this.middleName = middleName;
-		this.role = role;
-		this.login = login;
-		this.email = email;
-		this.status = status;
-		this.address = address;
-		this.passport = passport;
-	}
-	
-	public UserDTO(String firstName, String lastName, String middleName, String role, String login, String email,
-			String status, AddressDTO address, PassportDTO passport,String territorialCommunity,ResourceNumberDTOJSON resourceNumberDTOJSON) {
-		this.firstName = firstName;
-		this.lastName = lastName;
-		this.middleName = middleName;
-		this.role = role;
-		this.login = login;
-		this.email = email;
-		this.status = status;
-		this.address = address;
-		this.passport = passport;
-		this.territorialCommunity = territorialCommunity;
-		this.resourceNumberDTOJSON = resourceNumberDTOJSON;
-	}
+    public void setRole(String role) {
+        this.role = role;
+    }
 
-	public UserDTO() {
+    public String getStatus() {
+        return status;
+    }
 
-	}
+    public void setStatus(String status) {
+        this.status = status;
+    }
 
-	public String getFirstName() {
-		return firstName;
-	}
+    public ResourceNumberDTOJSON getResourceNumberDTOJSON() {
+        return resourceNumberDTOJSON;
+    }
 
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
+    public void setResourceNumberDTOJSON(ResourceNumberDTOJSON resourceNumberDTOJSON) {
+        this.resourceNumberDTOJSON = resourceNumberDTOJSON;
+    }
 
-	public String getLastName() {
-		return lastName;
-	}
+    public WillDocumentDTO getWillDocument() {
+        return willDocument;
+    }
 
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
+    public void setWillDocument(WillDocumentDTO willDocument) {
+        this.willDocument = willDocument;
+    }
 
-	public String getMiddleName() {
-		return middleName;
-	}
+    public List<String> getOtherDocuments() {
+        return otherDocuments;
+    }
 
-	public void setMiddleName(String middleName) {
-		this.middleName = middleName;
-	}
+    public void setOtherDocuments(List<String> otherDocuments) {
+        this.otherDocuments = otherDocuments;
+    }
 
-	public String getRole() {
-		return role;
-	}
-
-	public void setRole(String role) {
-		this.role = role;
-	}
-
-	public String getLogin() {
-		return login;
-	}
-
-	public void setLogin(String login) {
-		this.login = login;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getStatus() {
-		return status;
-	}
-
-	public void setStatus(String status) {
-		this.status = status;
-	}
-
-	public AddressDTO getAddress() {
-		return address;
-	}
-
-	public void setAddress(AddressDTO address) {
-		this.address = address;
-	}
-
-	public PassportDTO getPassport() {
-		return passport;
-	}
-
-	public void setPassport(PassportDTO passport) {
-		this.passport = passport;
-	}
-
-	public WillDocumentDTO getWillDocument() {
-		return willDocument;
-	}
-
-	public void setWillDocument(WillDocumentDTO willDocument) {
-		this.willDocument = willDocument;
-	}
-
-	public List<String> getOtherDocuments() {
-		return otherDocuments;
-	}
-
-	public void setOtherDocuments(List<String> otherDocuments) {
-		this.otherDocuments = otherDocuments;
-	}
-
-	@Override
-	public String toString() {
-		String result;
-		result = String.valueOf("ПІБ: " + lastName + " " + firstName + " " + middleName + "\n" + "Роль:  " + getRole()
-				+ "\n" + "Логін: " + login + "\n" + "Пошта: " + email + "\n" + "Статус: " + status + "\n"
-				+ "Паспортні дані:" + passport.toString() + "\n" + "Адреса: " + address.toString() + "\n");
-		return result;
-	}
-
-	public ResourceNumberDTOJSON getResourceNumberDTOJSON() {
-		return resourceNumberDTOJSON;
-	}
-
-	public void setResourceNumberDTOJSON(ResourceNumberDTOJSON resourceNumberDTOJSON) {
-		this.resourceNumberDTOJSON = resourceNumberDTOJSON;
-	}
-
-	public String getTerritorialCommunity() {
-		return territorialCommunity;
-	}
-
-	public void setTerritorialCommunity(String territorialCommunity) {
-		this.territorialCommunity = territorialCommunity;
-	}
 }
