+++ src/main/java/org/registrator/community/service/impl/UserServiceImpl.java
@@ -15,6 +15,7 @@ import org.registrator.community.dto.PassportDTO;
 import org.registrator.community.dto.ResourceNumberDTO;
 import org.registrator.community.dto.TomeDTO;
 import org.registrator.community.dto.UserDTO;
+import org.registrator.community.dto.UserRegistrationDTO;
 import org.registrator.community.dto.WillDocumentDTO;
 import org.registrator.community.dto.JSON.ResourceNumberDTOJSON;
 import org.registrator.community.dto.JSON.UserStatusDTOJSON;
@@ -29,7 +30,6 @@ import org.registrator.community.entity.User;
 import org.registrator.community.entity.WillDocument;
 import org.registrator.community.enumeration.RoleType;
 import org.registrator.community.enumeration.UserStatus;
-import org.registrator.community.forms.RegistrationForm;
 import org.registrator.community.service.CommunityService;
 import org.registrator.community.service.UserService;
 import org.slf4j.Logger;
@@ -408,6 +408,59 @@ public class UserServiceImpl implements UserService {
 	 */
 	@Override
 	@Transactional
+	public void registerUser(UserRegistrationDTO registrationForm) {
+	    TerritorialCommunity territorialCommunity = communityService
+	            .findByName(registrationForm.getTerritorialCommunity());
+	    User user = new User();
+	    user.setLogin(registrationForm.getLogin());
+	    user.setEmail(registrationForm.getEmail());
+	    user.setPasswordHash(userPasswordEncoder.encode(registrationForm.getPassword()));
+	    user.setFirstName(registrationForm.getFirstName());
+	    user.setLastName(registrationForm.getLastName());
+	    user.setMiddleName(registrationForm.getMiddleName());
+	    user.setPhoneNumber(registrationForm.getPhoneNumber());
+	    user.setRole(roleRepository.findRoleByType(RoleType.USER));
+	    user.setStatus(UserStatus.INACTIVE);
+	    user.setPhoneNumber(registrationForm.getPhoneNumber());
+	    user.setDateOfAccession(registrationForm.getDateOfAccession());
+	    user.setTerritorialCommunity(territorialCommunity);
+	    
+	    userRepository.save(user);
+	    log.info("Inserted new user data into 'users' table: user_id = " + user.getUserId());
+	    
+	    if (userRepository.findUserByLogin(user.getLogin()) != null) {
+	        // insert user's passport data into "passport_data" table
+	        PassportDTO passportDTO = registrationForm.getPassport();
+	        PassportInfo passport = new PassportInfo();
+	        passport.setUser(user);
+	        passport.setSeria(passportDTO.getSeria());
+	        passport.setNumber((passportDTO.getNumber()));
+	        passport.setPublishedByData(passportDTO.getPublished_by_data());
+	        passport.setComment(passportDTO.getComment());
+	        
+	        passportRepository.save(passport);
+	        log.info("Inserted passport data for user with passport_data_id", user.getLogin(),
+	                passport.getPassportId());
+	        
+	        // insert user's address records into "address" table
+	        AddressDTO addressDTO = registrationForm.getAddress();
+	        Address address = new Address();
+	        address.setUser(user);
+	        address.setCity(addressDTO.getCity());
+	        address.setRegion(addressDTO.getRegion());
+	        address.setDistrict(addressDTO.getDistrict());
+	        address.setStreet(addressDTO.getStreet());
+	        address.setBuilding(addressDTO.getBuilding());
+	        address.setFlat(addressDTO.getFlat());
+	        address.setPostCode(addressDTO.getPostcode());
+	        
+	        addressRepository.save(address);
+	        log.info("Inserted address data for user with address_id", user.getLogin(), address.getAddressId());
+	    }
+	    
+	}
+/*	@Override
+	@Transactional
 	public void registerUser(RegistrationForm registrationForm) {
 		TerritorialCommunity territorialCommunity = communityService
 				.findByName(registrationForm.getTerritorialCommunity());
@@ -456,7 +509,7 @@ public class UserServiceImpl implements UserService {
 		}
 
 	}
-
+*/
 	@Transactional
 	@Override
 	public boolean checkUsernameNotExistInDB(String login) {
