+++ src/main/java/org/registrator/community/validator/UserDataValidator.java
@@ -1,9 +1,7 @@
 package org.registrator.community.validator;
 
 import org.registrator.community.dao.UserRepository;
-import org.registrator.community.dto.ResourceTypeDTO;
-import org.registrator.community.dto.UserNameDTO;
-import org.registrator.community.forms.RegistrationForm;
+import org.registrator.community.dto.UserRegistrationDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
@@ -17,7 +15,7 @@ public class UserDataValidator implements Validator {
 
     @Override
     public boolean supports(Class<?> clazz) {
-        return clazz.equals(RegistrationForm.class);
+        return clazz.equals(UserRegistrationDTO.class);
     }
 
     /**
@@ -27,7 +25,7 @@ public class UserDataValidator implements Validator {
      */
     @Override
     public void validate(Object target, Errors errors) {
-        RegistrationForm registrationForm = (RegistrationForm) target;
+        UserRegistrationDTO registrationForm = (UserRegistrationDTO) target;
 
         if (userRepository.findUserByLogin(registrationForm.getLogin()) != null) {
             errors.rejectValue("login", "msg.registration.login.exist");
